wip: move EvmEvent to gstd, remove SysCalls
