+++ rs/macros/src/lib.rs
@@ -179,6 +179,57 @@ pub fn export(args: TokenStream, impl_item_fn_tokens: TokenStream) -> TokenStrea
     sails_macros_core::export(args.into(), impl_item_fn_tokens.into()).into()
 }
 
+/// Trait for encoding Ethereum events for the EVM.
+///
+/// This trait provides a uniform interface to convert an event into the topics and data payload
+/// that are used to emit logs in the Ethereum Virtual Machine (EVM). The logs generated by the EVM
+/// consist of:
+///
+/// - **Topics:** An array of 32-byte values. The first topic is always the keccak256 hash of the event
+///   signature, while the remaining topics correspond to indexed fields. For dynamic types (as determined
+///   by `<T as alloy_sol_types::SolType>::IS_DYNAMIC`), the ABI-encoded value is hashed before being stored.
+///   For static types, the ABI-encoded value is left-padded with zeros to 32 bytes.
+/// - **Data:** A byte array containing the ABI-encoded non-indexed fields of the event, encoded as a tuple.
+///
+/// This trait is intended to be used with the `#[derive(EvmEvent)]` procedural macro, which automatically
+/// implements the trait for your enum-based event definitions.
+///
+/// # Examples
+///
+/// Given an event definition:
+///
+/// ```rust
+/// #[derive(EvmEvent)]
+/// pub enum Events {
+///     MyEvent {
+///         #[indexed]
+///         sender: uint128,
+///         amount: uint128,
+///         note: String,
+///     },
+/// }
+/// ```
+///
+/// Calling the methods:
+///
+/// ```rust
+/// let event = Events::MyEvent {
+///     sender: 123,
+///     amount: 1000,
+///     note: "Hello, Ethereum".to_owned(),
+/// };
+///
+/// let topics = event.topics();
+/// let data = event.data();
+/// ```
+///
+/// The first topic will be the hash of the event signature (e.g. `"MyEvent(uint128,uint128,String)"`),
+/// and additional topics and the data payload will be computed based on the field attributes.
+///
+/// # Methods
+///
+/// - `topics()`: Returns a vector of 32-byte topics (`alloy_primitives::B256`) for the event.
+/// - `data()`: Returns the ABI-encoded data payload (a `Vec<u8>`) for the non-indexed fields.
 #[proc_macro_error]
 #[proc_macro_derive(EvmEvent, attributes(indexed, sails_path))]
 pub fn derive_evm_event(input: TokenStream) -> TokenStream {
