+++ rs/src/gstd/services.rs
@@ -54,6 +54,7 @@ impl ExposureContext {
 }
 
 #[cfg(not(feature = "ethexe"))]
+#[cfg(target_arch = "wasm32")]
 pub(crate) fn exposure_context(message_id: MessageId) -> ExposureContext {
     let map = get_message_id_to_service_route_map();
     let route = map
